Q working, sending to other services for feed
